refactor factor fact act t
